+++ includes/header.php
@@ -118,7 +118,7 @@
 </div>
     <nav class="navbar navbar-expand-lg navbar-custom">
         <div class="container">
-            <a class="navbar-brand" href="<?php echo getBaseUrl(); ?>">SHaDE</a>
+            <a class="navbar-brand" href="<?php echo getBaseUrl(); ?>">Home</a>
             <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>